BOAC-2289, to prep for batch-note tests, move 'create new note' elements to dedicated class
