+++ util/spec_helper.rb
@@ -76,6 +76,7 @@ require_relative '../pages/junction/canvas_mailing_lists_page'
 require_relative '../pages/junction/canvas_mailing_list_page'
 require_relative '../pages/junction/canvas_e_grades_export_page'
 require_relative '../pages/boac/boac_pages'
+require_relative '../pages/boac/boac_pages_create_note_modal'
 require_relative '../pages/boac/boac_group_modal_pages'
 require_relative '../pages/boac/boac_add_group_selector_pages'
 require_relative '../pages/boac/boac_list_view_pages'
